补充原理图

+++ bfstatus0.2-remote/Libraries/Lib/inc/stm32fx_delay.h
@@ -2,6 +2,6 @@
 #define __STM32FX_DELAY_H__
 #include "stm32f10x.h"
 void Init_SysTick(void);
-void DelayMs(__IO uint32_t nTime);
-void DelayUs(__IO uint32_t nTime);
+void DelayMs(__IO uint16_t nTime);
+void DelayUs(__IO uint16_t nTime);
 #endif
